# src/types/response/Doctor.Response.ts
export interface SaveUpdateResDoctor{
    _id:string,
    name:string,
    email:string,
    phone:string,
    specialist:string,
    avatar?:string,
    patient?:any

}

//when get record list of patient show 